[QA] 분석하기: 화살표 탭 이동 및 화면 비율 수정
